Frontend finished

+++ backend/app/Services/MovieService.php
@@ -14,15 +14,47 @@ use App\Http\Requests\UpdateMovieRequest;
 class MovieService
 {
 
-  public function getMovies(): array
+  public function getMovies(Request $request): array
   {
-    $movies = Movie::all();
-    return $movies->toArray();
+    $queries = $request->input('query');
+		$limit = $request->input('limit');
+		$ascending = $request->input('ascending');
+		$page = $request->input('page');
+		$orderBy = $request->input('orderBy');
+
+			$data = Movie::with('turns');
+
+		if (isset($queries) && $queries) {
+			foreach(json_decode($queries) as $field => $query) {
+				$data->where($field, 'LIKE', '%' . $query . '%');
+			};
+		}
+
+		$count = $data->count();
+		
+		if (isset($limit) && $limit) {
+      $data->limit($limit)->skip($limit * ($page - 1));
+    }
+				
+    if (isset($orderBy)) {
+      $direction = $ascending == 1 ? 'ASC' : 'DESC';
+    } else {
+      $direction = 'ASC';
+      $orderBy = 'id';
+    }
+    $data->orderBy($orderBy, $direction);
+
+		$results = $data->get()->toArray();
+
+		return [
+			'data'	=> $results,
+			'count' => $count
+		];
   }
 
   public function getMovie(int $id): ?Movie
   {
-    return Movie::find($id);
+    return Movie::with('turns')->where('id', $id)->first();
   }
 
   public function storeMovie(CreateMovieRequest $request): ?Movie
@@ -55,9 +87,11 @@ class MovieService
       
       if ($file = $request->file('image')) {
 
-        $exists = Storage::disk('local')->exists($oldImage);
-        if ($exists) {
-          Storage::delete($oldImage);
+        if ($oldImage) {
+          $exists = Storage::disk('local')->exists($oldImage);
+          if ($exists) {
+            Storage::delete($oldImage);
+          }
         }
 
         $path = $file->store('public/images');
